requirements.txt
Django==5.0.7
psycopg2-binary==2.9.9
django-phone-field==1.8.1
